Move context creator to a dedicated files
